refactor(web): plans page v1 deleted and replaced by v2
