--- apps/web/src/components/Plans/data.json
@@ -1,93 +1,0 @@
-{
-  "data": {
-    "monthly": [
-      {
-        "id": 1,
-        "weeklyMinutes": 120,
-        "price": 1800,
-        "label": "most-common"
-      },
-      {
-        "id": 2,
-        "weeklyMinutes": 180,
-        "price": 2700,
-        "primary": true
-      },
-      {
-        "id": 3,
-        "price": 3400,
-        "weeklyMinutes": 240,
-        "label": "most-valuable"
-      }
-    ],
-    "quarter": [
-      {
-        "id": 1,
-        "weeklyMinutes": 120,
-        "price": 1800,
-        "label": "most-common",
-        "discount": 5
-      },
-      {
-        "id": 2,
-        "weeklyMinutes": 180,
-        "price": 2700,
-        "discount": 7,
-        "primary": true
-      },
-      {
-        "id": 3,
-        "price": 3400,
-        "discount": 5,
-        "weeklyMinutes": 240,
-        "label": "most-valuable"
-      }
-    ],
-    "half": [
-      {
-        "id": 1,
-        "price": 1800,
-        "discount": 15,
-        "weeklyMinutes": 120,
-        "label": "most-common"
-      },
-      {
-        "id": 2,
-        "weeklyMinutes": 180,
-        "price": 2200,
-        "discount": 18,
-        "primary": true
-      },
-      {
-        "id": 3,
-        "price": 3400,
-        "discount": 15,
-        "weeklyMinutes": 240,
-        "label": "most-valuable"
-      }
-    ],
-    "annual": [
-      {
-        "id": 1,
-        "weeklyMinutes": 120,
-        "price": 1800,
-        "label": "most-common",
-        "discount": 30
-      },
-      {
-        "id": 2,
-        "weeklyMinutes": 180,
-        "price": 2700,
-        "discount": 35,
-        "primary": true
-      },
-      {
-        "id": 3,
-        "price": 3400,
-        "discount": 26,
-        "weeklyMinutes": 240,
-        "label": "most-valuable"
-      }
-    ]
-  }
-}
