+++ packages/ui/src/locales/ar-eg.json
@@ -748,8 +748,6 @@
   "plans.labels.annual": "سنوي",
   "plans.installments.three-month": "قسط علي 3 شهور",
   "plans.installments.six-month": "قسط علي 6 شهور",
-  "plans.subscribe.soon.title": "قريبا 🎉",
-  "plans.subscribe.soon.description": "بإمكانك التسجيل في الخطط بمجرد ان يتم تفعيل هذه الخاصية.",
   "plans.titles.beginning": "البداية",
   "plans.titles.advanced": "التقدم",
   "plans.titles.professional": "الاحتراف",
